move mock method out of production code
