+++ lib/open_project/files.rb
@@ -56,14 +56,5 @@ module OpenProject
       tmp = create_temp_file name: name, content: content, binary: binary
       Rack::Multipart::UploadedFile.new tmp.path, content_type, binary
     end
-
-    def mock_uploaded_file(name: 'test.txt',
-                           content_type: 'text/plain',
-                           content: 'test content',
-                           binary: false)
-
-      tmp = create_temp_file name: name, content: content, binary: binary
-      Rack::Test::UploadedFile.new tmp.path, content_type, binary
-    end
   end
 end
