+++ spec/legacy/unit/helpers/application_helper_spec.rb
@@ -47,10 +47,10 @@ describe ApplicationHelper, type: :helper do
 
     @issue = FactoryGirl.create :work_package, project: @project, author: @project_member, type: @project.types.first
 
-    file = mock_uploaded_file name: 'logo.gif',
-                                content_type: 'image/gif',
-                                content: 'not actually a gif',
-                                binary: true
+    file = FileHelpers.mock_uploaded_file name: 'logo.gif',
+                                          content_type: 'image/gif',
+                                          content: 'not actually a gif',
+                                          binary: true
     @attachment = FactoryGirl.create :attachment,
                                      author: @project_member,
                                      file: file,
