+++ spec/requests/api/v3/attachments/attachments_by_work_package_resource_spec.rb
@@ -32,7 +32,7 @@ require 'rack/test'
 describe 'API v3 Attachments by work package resource', type: :request do
   include Rack::Test::Methods
   include API::V3::Utilities::PathHelper
-  include OpenProject::Files
+  include FileHelpers
 
   let(:current_user) {
     FactoryGirl.create(:user,
